Fixed code smells

+++ src/include/insignia.h
@@ -20,7 +20,8 @@ static inline void cyancore_insignia()
 	extern char cyancore_ascii_insignia[];
 	extern ccver_t FWVersion;
 	printf("%s\nVersion %u.%u.%u\n", cyancore_ascii_insignia,
-		FWVersion.v.arch, FWVersion.v.major, FWVersion.v.minor);
+		(uint8_t) FWVersion.v.arch, (uint16_t) FWVersion.v.major,
+		(uint8_t) FWVersion.v.minor);
 }
 
 static inline void cyancore_insignia_lite()
@@ -28,5 +29,6 @@ static inline void cyancore_insignia_lite()
 	extern char cyancore_ascii_insignia_lite[];
 	extern ccver_t FWVersion;
 	printf("%s v%u.%u.%u\n", cyancore_ascii_insignia_lite,
-		FWVersion.v.arch, FWVersion.v.major, FWVersion.v.minor);
+		(uint8_t) FWVersion.v.arch, (uint16_t) FWVersion.v.major,
+		(uint8_t) FWVersion.v.minor);
 }
